packages/syftbox/pyproject.toml
[project]
name = "syftbox"
version = "0.3.5"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi==0.115.7",
    "uvicorn==0.34.0",
    "gunicorn==23.0.0",
    "jinja2==3.1.5",
    "typing-extensions==4.12.2",
    "pydantic-settings==2.7.1",
    "httpx==0.28.1",
    "pyyaml==6.0.2",
    "psutil==6.1.1",
    "loguru==0.7.3",
    "py-fast-rsync==0.1.0",
    "pathspec==0.12.1",
    "python-multipart==0.0.20",
    "rich==13.9.4",
    "croniter==6.0.0",
    "typer==0.15.1",
    "pid==3.0.4",
    "pydantic[email]==2.10.6",
    "pyjwt==2.10.1",
    "wcmatch==10.0",
    "curl-cffi>=0.7.4",
    "opentelemetry-instrumentation-fastapi==0.50b0",
    "opentelemetry-instrumentation-sqlite3==0.50b0",
    "opentelemetry-exporter-otlp-proto-grpc==1.29.0",
    "distro==1.9.0",
    "textual>=1.0.0",
    "msgpack>=1.1.0",
    "tqdm>=4.67.1",
    "aiofiles>=24.1.0",
    "requests>=2.32.3",
]

# Local dependencies for development
# Add using `uv add --group <group> <pip package>`
[dependency-groups]

# Published optional dependencies, or "extras". Will be referenced in the built wheel
# add using `uv add --optional <group> <pip package>`
[project.optional-dependencies]

[project.scripts]
syftbox = "syftbox.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
include = ["syftbox", "default_apps", "./uv.lock", "./README.md"]
exclude = ["syftbox/assets/icon/*"]

[tool.uv]
# add using `uv add --dev <group> <pip package>`
# this will be completely ignored in the built wheel
dev-dependencies = [
    "bump2version>=1.0.1",
    "faker>=30.3.5",
    "ipykernel>=6.29.5",
    "locust>=2.32.0",
    "pre-commit>=4.0.1",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "pytest-httpx>=0.35.0",
    "pytest-timeout>=2.3.1",
    "pytest-xdist[psutil]>=3.6.1",
    "pytest>=8.3.3",
    "textual-dev>=1.7.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 120
exclude = [".venv", "dist", ".clients", ".e2e", ".server"]

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.lint.per-file-ignores]
"**/__init__.py" = ["F401"]

[tool.coverage.report]
skip_empty = true

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
'syftbox' = ['syftbox/server2client_version.json']
